Move staking contracts to TerraDotnet library
