+++ test/NarrowIntegrationTests/Oracles/TerraLcdClient/TerraMoneyFcdApiClientTests.cs
@@ -3,6 +3,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Pylonboard.Infrastructure.Hosting.TerraDataFetchers;
 using Refit;
+using TerraDotnet;
 using TerraDotnet.TerraFcd;
 using Xunit;
 using Xunit.Abstractions;
